+++ docs/changelog.rst
@@ -14,6 +14,8 @@ Under development
 
 * Improvement: Added support for variants handling for need options.
   (`#671 <https://github.com/useblocks/sphinx-needs/issues/671>`_)
+* Improvement: Added Jinja support for need content via the :ref:`jinja_content` option.
+  (`#678 <https://github.com/useblocks/sphinx-needs/issues/678>`_)
 * Improvement: Added checks and warnings for :ref:`needimport` and :ref:`needs_external_needs`.
   (`#624 <https://github.com/useblocks/sphinx-needs/issues/624>`_)
 * Improvement: Support for :ref:`needs_string_links` in :ref:`needtable`.